Refactor JettyController to use namespaced views, update routes for consistency, and enhance inventory models with new attributes and relationships. Introduce methods for quantity validation in JettyMap and update inventory table view to display expiration dates. Remove deprecated gear seeder and adjust inventory seeder for improved variant retrieval.

+++ database/migrations/2025_07_23_123205_create_inventory_variant_table.php
@@ -30,8 +30,8 @@ return new class extends Migration
             $table->id();
             $table->enum('type', self::INVENTORY_TYPES);
             $table->string('name');
-            $table->decimal('swl', 8, 2)->default(0);
-            $table->decimal('length', 8, 2)->default(0);
+            $table->float('swl', 2)->default(0);
+            $table->float('length', 2)->default(0);
             $table->timestamps();
 
             $table->unique(['type', 'name', 'swl', 'length'], 'type_name_swl_length_unique');
